update function header error

+++ include/secp256k1_bulletproofs.h
@@ -126,7 +126,7 @@ SECP256K1_WARN_UNUSED_RESULT SECP256K1_API int secp256k1_bulletproof_rangeproof_
     unsigned char* blind,
     const unsigned char* proof,
     size_t plen,
-    size_t min_value,
+    uint64_t min_value,
     const secp256k1_pedersen_commitment* commit,
     const secp256k1_generator* value_gen,
     const unsigned char* nonce,
